Migrate to spring 4.3.x and transitive dependencies (ehcache, javassist, hibernate, mybatis)
- Fixup minor breakages because of API changes in upgrades of libraries
- Use a different `CacheManager` instances for `goCache`,
  `CachingSubjectDnX509PrincipalExtractor`, and for hibernate

+++ server/src/main/java/com/thoughtworks/go/server/dao/UserDao.java
@@ -31,7 +31,7 @@ public interface UserDao {
 
     Users allUsers();
 
-    Integer enabledUserCount();
+    long enabledUserCount();
 
     void deleteAll();
